[Malterlib/Container] Fix incorrect stream of registry keys

+++ Source/Malterlib_Container_Registry_Key.hpp
@@ -337,7 +337,7 @@ namespace NMib::NContainer
 			_Stream >> m_Location;
 			_Stream >> m_ValueLocation;
 
-			uint8 Flags;
+			uint32 Flags;
 			_Stream >> Flags;
 			m_bForceEscapedKey = Flags & 1u;
 			m_bForceEscapedValue = (Flags >> 1u) & 1u;
